requirements.txt
certifi==2019.3.9
chardet==3.0.4
cycler==0.10.0
dj-database-url==0.5.0
Django==2.1.7
django-cors-headers==2.5.2
django-heroku==0.3.1
djangorestframework==3.9.2
idna==2.8
kiwisolver==1.0.1
matplotlib==3.0.3
numpy==1.16.2
oauthlib==3.0.1
pandas==0.24.2
psycopg2==2.7.7
pyparsing==2.3.1
PySocks==1.6.8
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
tweepy==3.7.0
urllib3==1.24.1
whitenoise==4.1.2
